deprecation_module: Move to new internal_test_util directory.
Now we no longer need to mess with sys.path in deprecation_test.
