+++ tests/BUILD
@@ -473,10 +473,10 @@ py_test(
 py_test(
     name = "deprecation_test",
     srcs = [
-        "deprecation_module/__init__.py",
         "deprecation_test.py",
     ],
     deps = [
+        "//jax:internal_test_util",
         "//jax:test_util",
     ],
 )
